[REFACTOR] #138: interior, exterior 카드 컴포넌트 분리
