+++ frontend/src/containers/ModelTypePage/ModelBannerContainer.tsx
@@ -3,7 +3,6 @@ import { BodyKrRegular4, BodyKrRegular5, HeadingKrRegular2 } from '../../styles/
 import CenterWrapper from '../../components/layout/CenterWrapper';
 import Banner from '../../components/common/banner/Banner';
 import HmgTag from '../../components/common/hmgTag/HmgTag';
-import PriceStaticBar from '../../components/priceStaticBar/PriceStaticBar';
 
 export default function ModelBannerContainer() {
   return (
